added eslint support and some style refactors also modified the structure to have .feature files in different directory
