+++ tests/frontend/test_kibana_app/cypress.json
@@ -1,6 +1,5 @@
 {
   "testFiles": "**/*.{feature,features}",
-  "ignoreTestFiles": ["*.js", "*.ts", "*.md"],
   "viewportWidth": 1360,
   "viewportHeight": 790
 }
