--- tests/frontend/test_kibana_app/cypress/integration/features/management/decoders/ShouldListDecodersAndShowThePager/common.spec.js
@@ -1,8 +1,0 @@
-import {When} from "cypress-cucumber-preprocessor/steps";
-import WzMenu from "../../../../pageobjects/wzMenu/WzMenu";
-
-When('The user navigates to decoders', () => {
-    const wzMenu = new WzMenu();
-    wzMenu.goToDecoders();
-    cy.wait(3000);
-});
