+++ tests/frontend/test_kibana_app/cypress/integration/features/management/decoders/should-list-decoders-and-show-the-pager.feature
