+++ tests/frontend/test_kibana_app/cypress/integration/features/management/rules/should-list-rules-and-show-the-pager.feature
