--- tests/frontend/test_kibana_app/cypress/integration/pageobjects/login/ODEFLoginPage.js
@@ -1,7 +1,0 @@
-class ODEFLoginPage {
-    static inputUsernameSelector = 'input[data-test-subj="user-name"]';
-    static inputPasswordSelector = 'input[data-test-subj="password"]';
-    static buttonSubmitSelector = 'button[data-test-subj="submit"]';
-}
-
-export default ODEFLoginPage;
