--- tests/frontend/test_kibana_app/cypress/integration/pageobjects/login/XPackLoginPage.js
@@ -1,7 +1,0 @@
-class XPackLoginPage {
-    static inputUsernameSelector = 'input[data-test-subj="loginUsername"]';
-    static inputPasswordSelector = 'input[data-test-subj="loginPassword"]';
-    static buttonSubmitSelector = 'button[data-test-subj="loginSubmit"]';
-}
-
-export default XPackLoginPage;
