+++ tests/frontend/test_kibana_app/cypress/integration/pageobjects/login/odef-login-page.js
@@ -1,0 +1,7 @@
+class OdefLoginPage {
+  static inputUsernameSelector = 'input[data-test-subj="user-name"]';
+  static inputPasswordSelector = 'input[data-test-subj="password"]';
+  static buttonSubmitSelector = 'button[data-test-subj="submit"]';
+}
+
+export default OdefLoginPage;
