+++ tests/frontend/test_kibana_app/cypress/integration/pageobjects/login/xpack-login-page.js
@@ -1,0 +1,7 @@
+class XpackLoginPage {
+  static inputUsernameSelector = 'input[data-test-subj="loginUsername"]';
+  static inputPasswordSelector = 'input[data-test-subj="loginPassword"]';
+  static buttonSubmitSelector = 'button[data-test-subj="loginSubmit"]';
+}
+
+export default XpackLoginPage;
