+++ tests/frontend/test_kibana_app/cypress/integration/pageobjects/wzMenu/rules.js
@@ -1,0 +1,42 @@
+import { getObject } from '../../utils/driver';
+
+class Rules {
+
+  constructor() {
+    this.rulesButtonSelector = '[class="euiSideNavItem euiSideNavItem--trunk"]';
+    this.titleSelector = '[class="euiTitle euiTitle--medium"]';
+    this.tableSelector = '[class="euiTableRow customRowClass euiTableRow-isClickable"]';
+    this.dropdownPaginationSelector = '[data-test-subj="tablePaginationPopoverButton"]';
+    this.listPagesSelector = 'nav[class="euiPagination"]';
+    this.customRulesButtonSelector = '[class="euiButtonGroup euiButtonGroup--m"]';
+
+  }
+
+  getRulesButton() {
+    return getObject(this.rulesButtonSelector)
+      .eq(0);
+  }
+
+  getTitle() {
+    return getObject(this.titleSelector);
+  }
+
+  getTable() {
+    return getObject(this.tableSelector);
+  }
+
+  getDropdownPagination() {
+    return getObject(this.dropdownPaginationSelector);
+  }
+
+  getListPages() {
+    return getObject(this.listPagesSelector);
+  }
+
+  getCustomRulesButton() {
+    return getObject(this.customRulesButtonSelector);
+  }
+
+}
+
+export default Rules;
