+++ tests/frontend/test_kibana_app/cypress/integration/step_definitions/management/decoders/common.js
@@ -1,0 +1,8 @@
+import { When } from 'cypress-cucumber-preprocessor/steps';
+import WazuhMenu from '../../../pageobjects/wzMenu/wazuh-menu';
+
+When('The user navigates to decoders', () => {
+  const wzMenu = new WazuhMenu();
+  wzMenu.goToDecoders();
+  cy.wait(3000);
+});
