+++ tests/frontend/test_kibana_app/cypress/integration/step_definitions/management/decoders/shouldCanEditADecoder.js
@@ -1,0 +1,23 @@
+import { Then, When } from 'cypress-cucumber-preprocessor/steps';
+import WazuhMenu from '../../../pageobjects/wzMenu/wazuh-menu';
+
+const wzMenu = new WazuhMenu();
+const decoders = wzMenu.getDecoders();
+
+When('The user presses the edit decoder button and edits it', () => {
+  cy.wait(3000);
+  decoders.getManageDecodersFilesButton()
+    .click();
+  decoders.getEditDecoderButton()
+    .click();
+  decoders.getSaveDecoderButton()
+    .click();
+});
+
+Then('The user should see the message', () => {
+  decoders.getMessageConfirmSave()
+    .should('have.text', 'Changes will not take effect until a restart is performed.');
+  decoders.getButtonRestart()
+    .should('exist')
+    .should('be.visible');
+});
