+++ tests/frontend/test_kibana_app/cypress/integration/step_definitions/management/rules/common.js
@@ -1,0 +1,9 @@
+import { When } from 'cypress-cucumber-preprocessor/steps';
+
+import WazuhMenu from '../../../pageobjects/wzMenu/wazuh-menu';
+
+When('The user navigates to rules', () => {
+  const wzMenu = new WazuhMenu();
+  wzMenu.goToRules();
+  cy.wait(3000);
+});
