+++ tests/frontend/test_kibana_app/cypress/plugins/index.js
@@ -22,5 +22,5 @@ const cucumber = require('cypress-cucumber-preprocessor').default
 module.exports = (on, config) => {
   // `on` is used to hook into various events Cypress emits
   // `config` is the resolved Cypress config
-    on('file:preprocessor', cucumber())
-}
+  on('file:preprocessor', cucumber());
+};
